Use custom ECs
